fix!(api): fixed gemeni api setup, needs props
