+++ src/app/(routes)/test/page.tsx
@@ -1,53 +1,75 @@
 "use client";
 import React, { useState } from "react";
 import { FileUploader } from "react-drag-drop-files";
+import { GoogleAIFileManager } from "@google/generative-ai/server";
+import { GoogleGenerativeAI } from "@google/generative-ai";
+import path from "path";
+import Gemeni from "../../api/gemeniapi";
+
 
 const fileTypes = ["JPG", "PNG", "GIF"];
 
 const Test = () => {
   const [file, setFile] = useState<File | null>(null);
-  const [response, setResponse] = useState<string | null>(null);
+  const [response, setResponse] = useState('');
 
-  const handleChange = (selectedFile: File) => {
-    setFile(selectedFile);
-  };
 
   const handleUpload = async () => {
-    if (!file) {
-      console.log("No file selected.");
-      return;
-    }
+    const someResponse = await Gemeni();   
+    setResponse(someResponse || '');
+  }
 
-    const formData = new FormData();
-    formData.append("fileData", file);
-    formData.append("mimeType", file.type);
-    formData.append("displayName", file.name);
+  //   if (!file) {
+  //     console.log("No file selected.");
+  //     return;
+  //   }
 
-    const res = await fetch("/api/gemeni", {
-      method: "POST",
-      body: formData,
-    });
+  //   const formData = new FormData();
+  //   formData.append("fileData", file);
 
-    const data = await res.json();
-    if (res.ok) {
-      setResponse(data.response);
-    } else {
-      console.log("Error from API:");
-    }
-  };
+  //   // Debugging FormData
+  //   console.log(formData.get("fileData"));
+
+  //   try {
+  //     const res = await fetch("/api/gemeni", {
+  //       method: "POST",
+  //       body: formData,
+  //     });
+
+  //     if (!res.ok) {
+  //       const errorText = await res.text();  // Capture error message
+  //       console.log("Error from API:", res.status, errorText);
+  //       return;
+  //     }
+
+  //     // Process the response as a Blob (image or file)
+  //     const blob = await res.blob();
+  //     const url = URL.createObjectURL(blob);
+
+  //     // Create a link to download the returned file
+  //     const link = document.createElement("a");
+  //     link.href = url;
+  //     link.download = file.name; // Use the original file name or set a custom one
+  //     link.click();
+  //     URL.revokeObjectURL(url);
+
+  //     console.log("File downloaded successfully.");
+  //   } catch (error) {
+  //     console.error("File upload failed:", error);
+  //   }
+  // };
 
   return (
     <div>
       <div className="p-20">
         <FileUploader
-            multiple={false}
-            handleChange={(e: any) => setFile(e)}
-            types={fileTypes}
-            name="file"
-            label="Upload or Drop a File"
-            required
-          />
-
+          multiple={false}
+          handleChange={(file: File) => setFile(file)}
+          types={fileTypes}
+          name="file"
+          label="Upload or Drop a File"
+          required
+        />
       </div>
       <button
         className="ml-20 p-20 text-white bg-black rounded-xl"
@@ -55,7 +77,7 @@ const Test = () => {
       >
         Upload and Process
       </button>
-      {response ? <p>Response: {response}</p> : null}
+      {response ? <p>{response}</p> : null}
     </div>
   );
 };
